seperated alertText function into its own component
