+++ blogproject/src/ViewDrafts.js
@@ -3,14 +3,13 @@ import {Link} from 'react-router-dom'
 import dayjs from 'dayjs'
 import AdvancedFormat from 'dayjs/plugin/advancedFormat'
 // import AdvancedFormat from 'dayjs/plugin/advancedFormat' // ES 2015
-
+import AlertText from './AlertText'
 import TimePicker from 'react-time-picker'
 import Container from 'react-bootstrap/Container'
 import Table from 'react-bootstrap/Table'
 import Row from 'react-bootstrap/Row'
 import ButtonGroup from 'react-bootstrap/ButtonGroup'
 import Dropdown from 'react-bootstrap/Dropdown'
-import Alert from 'react-bootstrap/Alert'
 
 import Form from 'react-bootstrap/Form'
 import Button from 'react-bootstrap/Button'
@@ -24,8 +23,8 @@ const [displayId, setDisplayId] = useState(null)
 
 const [showFilter, toggleFilter] = useState(null)
 const [showSort, toggleSort] = useState(null)
-const [showAlert, setShowAlert] = useState(true);
-const [AlerMessage, setAlertMessage] = useState(true);
+const [showAlert, setShowAlert] = useState(false);
+const [AlertMessage, setAlertMessage] = useState(true);
 
 const [beginningDate, setBeginningDate] = useState(null)
 const [endingDate, setEndingDate] = useState(null)
@@ -118,7 +117,7 @@ useEffect(()=>{
 
     return (
         <Container fluid style={{border: '2px dashed red', marginBottom: '2rem'}}>
-        <AlertText  showAlert={showAlert} setShowAlert={setShowAlert} AlerMessage={AlerMessage}/>
+        <AlertText  showAlert={showAlert} setShowAlert={setShowAlert} AlertMessage={AlertMessage}/>
 
 
         <SearchOptions array={mappedArray} setNewArray={setMappedArray}  setAlertMessage={setAlertMessage} setShowAlert={setShowAlert}/>
@@ -501,19 +500,6 @@ if( !searchTerm ){
     )
 }
 
-const AlertText =(props) => {
-  
-    if (props.showAlert) {
-      return (
-        <Alert variant="danger" onClose={() => props.setShowAlert(false)} dismissible>
-          <Alert.Heading>Oh snap!</Alert.Heading>
-          <p>
-            {props.AlerMessage}
-          </p>
-        </Alert>
-      );
-    }
-    return <Button onClick={() => props.setShowAlert(true)}>Show Alert</Button>;
-  }
+
   
 export default ViewDrafts
